requirements.txt
contourpy==1.3.3
cycler==0.12.1
fonttools==4.61.1
ia870==0.8
ImageIO==2.37.2
kiwisolver==1.4.9
lazy-loader==0.5
libcst==1.8.6
matplotlib==3.10.8
networkx==3.6.1
numpy==2.4.2
opencv-python==4.13.0.92
packaging==26.0
pillow==12.1.1
pyparsing==3.3.2
python-dateutil==2.9.0.post0
PyWavelets==1.9.0
PyYAML==6.0.3
ruff==0.15.5
scikit-image==0.26.0
scipy==1.17.1
six==1.17.0
tifffile==2026.3.3
